feat: Reorganize code with modules
