+++ src/components/EmptyState/assets/bug-illustration.png
