+++ src/components/EmptyState/assets/wip-illustration.png
