+++ src/stories/index.js
@@ -3,8 +3,8 @@ import './badge/BadgeVariants.stories';
 import './button/ButtonColors.stories';
 import './button/ButtonSizes.stories';
 import './button/ButtonVariants.stories';
-import './gridview/GridViewStandard.stories';
-import './layout/EmptyStateAction.stories';
-import './layout/EmptyStateStandard.stories';
+import './empty-state/EmptyStateAction.stories';
+import './empty-state/EmptyStateStandard.stories';
+import './grid-view/GridViewStandard.stories';
 import './styleguide/Colors.stories';
 import './toast/ToastStatuses.stories';
