+++ src/stories/toast/ToastStatuses.stories.js
@@ -4,8 +4,8 @@ import { storiesOf } from '@storybook/react-native';
 import { Button } from 'native-base';
 
 import { useToast } from '@/components/Toast';
-import StoryLayout from '@/screens/Dev/Storybook/_partials/StoryLayout';
-import StorySpacer from '@/screens/Dev/Storybook/_partials/StorySpacer';
+import StoryLayout from '@/devtools/storybook/_partials/StoryLayout';
+import StorySpacer from '@/devtools/storybook/_partials/StorySpacer';
 
 const toastStory = storiesOf('Toast', module);
 const storyTitle = 'Toast statuses';
