+++ src/tests/App.test.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import 'react-native';
+// Note: test renderer must be required after react-native.
+import renderer from 'react-test-renderer';
+
+import App from '@/App';
+
+it('renders correctly', () => {
+  renderer.create(<App />);
+});
